date format_2

+++ app/src/main/java/com/example/photosdemo/data/models/image/ImageDtoIn.kt
@@ -2,7 +2,7 @@ package com.example.photosdemo.data.models.image
 
 data class ImageDtoIn(
     var base64Image: String?,
-    var date: Int?,
+    var date: Long?,
     var lat: Double?,
     var lng: Double?
 )
